refactor: UIViewファイルへViewを切り出し
